[Feat] scrollDismissesKeyboard 추가 및 파일 분리
